Update 05. Doubly Circulary List.cpp

+++ 05. Doubly Circulary List.cpp
@@ -10,10 +10,10 @@ fstream file("Student.txt", ios::in);
 class Student
 {
 public:
-	int id;
+	size_t id;
 	char* name;
 	float average;
-	int sizeMarks;
+	size_t sizeMarks;
 	int* marks;
 
 	Student()
@@ -73,7 +73,7 @@ void readFromFile(Student &s)
 	if (s.sizeMarks > 0)
 	{
 		s.marks = new int[s.sizeMarks];
-		for (int i = 0; i < s.sizeMarks; i++)
+		for (size_t i = 0; i < s.sizeMarks; i++)
 		{
 			file >> s.marks[i];
 		}
@@ -109,7 +109,7 @@ void display(DoublyCircularyList* list, DoublyCircularyList* endList)
 	cout << "\nAverage: " << list->s.average;
 	cout << "\nNumber of marks: " << list->s.sizeMarks;
 	cout << "\nMarks: ";
-	for (int i = 0; i < list->s.sizeMarks; i++)
+	for (size_t i = 0; i < list->s.sizeMarks; i++)
 	{
 		cout << list->s.marks[i] << " ";
 	}
